+++ lib/vagrant/provisioners/puppet_server.rb
@@ -21,8 +21,8 @@ module Vagrant
       end
 
       def provision!
-        verify_binary("puppetd")
-        run_puppetd_client
+        verify_binary("puppet")
+        run_puppet_agent
       end
 
       def verify_binary(binary)
@@ -32,7 +32,7 @@ module Vagrant
                               :binary => binary)
       end
 
-      def run_puppetd_client
+      def run_puppet_agent
         options = config.options
         options = [options] if !options.is_a?(Array)
 
@@ -66,7 +66,7 @@ module Vagrant
           facter = "#{facts.join(" ")} "
         end
 
-        command = "#{facter}puppetd #{options} --server #{config.puppet_server} --detailed-exitcodes || [ $? -eq 2 ]"
+        command = "#{facter}puppet agent #{options} --server #{config.puppet_server} --detailed-exitcodes || [ $? -eq 2 ]"
 
         env[:ui].info I18n.t("vagrant.provisioners.puppet_server.running_puppetd")
         env[:vm].channel.sudo(command) do |type, data|